move snapshotter config structs to implementation
